--- app/__init__.py
@@ -1,1 +1,0 @@
-from app.app import *
