+++ app/route/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+main = Blueprint("main", __name__)
+
+from . import create, index, login, signup, thread  # noqa: E402
