+++ app/route/index.py
@@ -1,0 +1,8 @@
+from flask import render_template
+
+from . import main
+
+
+@main.route("/", methods=["POST", "GET"])
+def index():
+    return render_template("index.html")
