+++ app/route/login.py
@@ -1,6 +1,8 @@
 from flask import render_template
 
-def login_route(app):
-    @app.route('/login', methods=["POST","GET"])
-    def login():
-        return render_template('login.html')
+from . import main
+
+
+@main.route("/login", methods=["POST", "GET"])
+def login():
+    return render_template("login.html")
