+++ app/route/signup.py
@@ -1,20 +1,20 @@
-from flask import Flask,request, jsonify
-from flask_sqlalchemy import SQLAlchemy
-from uuid import uuid4
-from app import User
-import datetime
+from flask import jsonify, render_template, request
+from models import User, db
 
-def signup_route(app, db):
-    @app.route('/signup', methods=['POST'])
-    def signup():
-        id = str(uuid4())
-        name = request.form.get('name')
-        email = request.form.get('email')
-        password = request.form.get('password')
-        created_at = datetime.now()
+from . import main
 
-        new_user = User(id=id, name=name, email=email, password=password, created_at=created_at)
-        db.session.add(new_user)
-        db.session.commit()
 
-        return jsonify({'message': 'User created successfully'}), 201
+@main.route("/signup", methods=["POST", "GET"])
+def signup():
+    if request.method == "GET":
+        return render_template("signup.html")
+
+    name = request.form.get("name")
+    email = request.form.get("email")
+    password = request.form.get("password")
+
+    new_user = User(name=name, email=email, password=password)
+    db.session.add(new_user)
+    db.session.commit()
+
+    return jsonify({"message": "User created successfully"}), 201
