+++ app/route/thread.py
@@ -1,6 +1,8 @@
 from flask import render_template
 
-def thread_route(app):
-    @app.route("/thread", methods=["POST","GET"])
-    def thread():
-        return render_template('thread.html')
+from . import main
+
+
+@main.route("/thread", methods=["POST", "GET"])
+def thread():
+    return render_template("thread.html")
