+++ app/templates/signup.html
@@ -16,17 +16,19 @@
         <form method="POST" class="space-y-3">
             <div>
                 <label for="" class="block text-sm font-midium text-gray-700">ユーザー名</label>
-                <input type="text" name="username" class="mt-1 block w-full py-2 px-3 border border-gray-300 rounded-md">
+                <input type="text" name="name" class="mt-1 block w-full py-2 px-3 border border-gray-300 rounded-md">
             </div>
 
             <div>
                 <label class="block text-sm font-medium text-gray-700" for="">メールアドレス</label>
-                <input type="email" name="email" class="mt-1 block w-full py-2 px-3 border border-gray-300 rounded-md" required>
+                <input type="email" name="email" class="mt-1 block w-full py-2 px-3 border border-gray-300 rounded-md"
+                    required>
             </div>
 
             <div>
                 <label class="block text-sm font-medium text-gray-700" for="">パスワード</label>
-                <input type="password" name="password" class="mt-1 block w-full py-2 px-3 border border-gray-300 rounded-md" required>
+                <input type="password" name="password"
+                    class="mt-1 block w-full py-2 px-3 border border-gray-300 rounded-md" required>
             </div>
 
             <div>
@@ -36,4 +38,5 @@
         </form>
     </div>
 </body>
+
 </html>
